refactor(demo): quarantine roles into DemoModule (ADR-0001)
First move of the one-way demo->IdP boundary: roles/ now lives under
demo/ so app-level role storage is no longer part of the IdP surface.
