+++ api/src/demo/demo.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { RolesModule } from './roles/roles.module';
+
+@Module({
+  imports: [RolesModule],
+  exports: [RolesModule],
+})
+export class DemoModule {}
