+++ api/src/guards/app-user.guard.ts
@@ -1,5 +1,5 @@
 import { Injectable, CanActivate, ExecutionContext, ForbiddenException, Logger } from '@nestjs/common';
-import { RolesService } from '../roles/roles.service';
+import { RolesService } from '../demo/roles/roles.service';
 
 @Injectable()
 export class AppUserGuard implements CanActivate {
